+++ db/container_test.go
@@ -30,14 +30,13 @@ func TestContainerString(t *testing.T) {
 		Status:     "testing",
 		Hostname:   "hostname",
 		Command:    []string{"run", "/bin/sh"},
-		Labels:     []string{"label1"},
 		Env:        fakeMap,
 		Created:    fakeTime,
 	}
 
 	exp = "Container-1{run test/test run /bin/sh, ImageID: imageid, " +
 		"DockerID: DockerID, Minion: Test, StitchID: 1, IP: 1.2.3.4, " +
-		"Hostname: hostname, Labels: [label1], Env: map[test:tester], " +
+		"Hostname: hostname, Env: map[test:tester], " +
 		"Status: testing, Created: " + fakeTimeString + "}"
 
 	assert.Equal(t, exp, c.String())